renamed core sender and receiver
