+++ pkg/node.go
@@ -1,0 +1,3 @@
+package giga
+
+// defines the senders and receivers for other packages to import?
